move: storybook 구조 변경
