+++ components/src/lib/storybook/Button.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import Button from '../Button';
+import Button from '../Button/Button';
 
 const meta = {
   title: 'Components/Button',
